fix: Removed remaining reference to i1.png
When initialising CUDA, the code reads an image to obtain the size. This contains
a reference to i1.png which had not been replaced by a reference to the provided
command line argument.  When i1.png is missing, the code fails silently and causes
the output pdflow files to be zero valued.

This change forces the code to read the first provided image parameter.

It doesn't fix the silent fail issue but a missing input file will be caught later
when processing occurs.